replymate-ai/requirements.txt
# === 7. requirements.txt ===
Flask
openai
twilio
psycopg2-binary
python-dotenv